chore(api): Update submodule to new commit
This commit updates the `api` submodule pointer from commit `07cc9f8` to `529a7a6`.

The purpose of this update is to incorporate the latest changes from the API repository into the main project.

Note: The new commit is marked as `dirty`, which indicates there are uncommitted changes in the submodule's working directory.